Move IntraAgentMessageUtils into graph.activations as an implementation detail there
